requirements-unit.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --resolver=backtracking ./requirements-unit.in
#
anyio==3.6.2
    # via
    #   -r ./requirements.txt
    #   httpcore
attrs==22.1.0
    # via pytest
cachetools==5.2.0
    # via
    #   -r ./requirements.txt
    #   google-auth
certifi==2022.9.24
    # via
    #   -r ./requirements.txt
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
charset-normalizer==2.1.1
    # via
    #   -r ./requirements.txt
    #   requests
coverage==6.5.0
    # via -r ./requirements-unit.in
exceptiongroup==1.0.4
    # via pytest
google-auth==2.15.0
    # via
    #   -r ./requirements.txt
    #   kubernetes
h11==0.14.0
    # via
    #   -r ./requirements.txt
    #   httpcore
httpcore==0.16.2
    # via
    #   -r ./requirements.txt
    #   httpx
httpx==0.23.1
    # via
    #   -r ./requirements.txt
    #   lightkube
idna==3.4
    # via
    #   -r ./requirements.txt
    #   anyio
    #   requests
    #   rfc3986
iniconfig==1.1.1
    # via pytest
jinja2==3.1.2
    # via -r ./requirements.txt
kubernetes==25.3.0
    # via -r ./requirements.txt
lightkube==0.11.0
    # via -r ./requirements.txt
lightkube-models==1.25.4.4
    # via
    #   -r ./requirements.txt
    #   lightkube
markupsafe==2.1.1
    # via
    #   -r ./requirements.txt
    #   jinja2
oauthlib==3.2.2
    # via
    #   -r ./requirements.txt
    #   requests-oauthlib
oci-image==1.0.0
    # via -r ./requirements.txt
ops==1.5.4
    # via -r ./requirements.txt
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
pyasn1==0.4.8
    # via
    #   -r ./requirements.txt
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   -r ./requirements.txt
    #   google-auth
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via
    #   -r ./requirements-unit.in
    #   pytest-lazy-fixture
    #   pytest-mock
pytest-lazy-fixture==0.6.3
    # via -r ./requirements-unit.in
pytest-mock==3.10.0
    # via -r ./requirements-unit.in
python-dateutil==2.8.2
    # via
    #   -r ./requirements.txt
    #   kubernetes
pyyaml==6.0
    # via
    #   -r ./requirements.txt
    #   kubernetes
    #   lightkube
    #   ops
requests==2.28.1
    # via
    #   -r ./requirements.txt
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   -r ./requirements.txt
    #   kubernetes
rfc3986[idna2008]==1.5.0
    # via
    #   -r ./requirements.txt
    #   httpx
rsa==4.9
    # via
    #   -r ./requirements.txt
    #   google-auth
six==1.16.0
    # via
    #   -r ./requirements.txt
    #   google-auth
    #   kubernetes
    #   python-dateutil
sniffio==1.3.0
    # via
    #   -r ./requirements.txt
    #   anyio
    #   httpcore
    #   httpx
tenacity==8.1.0
    # via -r ./requirements.txt
tomli==2.0.1
    # via pytest
urllib3==1.26.13
    # via
    #   -r ./requirements.txt
    #   kubernetes
    #   requests
websocket-client==1.4.2
    # via
    #   -r ./requirements.txt
    #   kubernetes

# The following packages are considered to be unsafe in a requirements file:
# setuptools
